+++ src/crawlee/playwright_crawler/playwright_crawler.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+import logging
 from typing import TYPE_CHECKING, Literal
 
 from typing_extensions import Unpack
@@ -52,6 +53,8 @@ class PlaywrightCrawler(BasicCrawler[PlaywrightCrawlingContext]):
         kwargs['_context_pipeline'] = ContextPipeline().compose(self._page_goto)
         kwargs['_additional_context_managers'] = [self._browser_pool]
 
+        kwargs.setdefault('_logger', logging.getLogger(__name__))
+
         super().__init__(**kwargs)
 
     async def _page_goto(
@@ -100,6 +103,7 @@ class PlaywrightCrawler(BasicCrawler[PlaywrightCrawlingContext]):
             send_request=context.send_request,
             push_data=context.push_data,
             proxy_info=context.proxy_info,
+            log=context.log,
             page=crawlee_page.page,
             enqueue_links=enqueue_links,
         )