+++ tests/unit/playwright_crawler/test_playwright_crawler.py
@@ -7,6 +7,7 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 from unittest import mock
 
+from crawlee import Glob
 from crawlee.playwright_crawler import PlaywrightCrawler
 
 if TYPE_CHECKING:
@@ -34,34 +35,18 @@ async def test_basic_request(httpbin: str) -> None:
 
 
 async def test_enqueue_links() -> None:
-    requests = ['https://crawlee.dev/']
+    requests = ['https://crawlee.dev/docs/examples']
     crawler = PlaywrightCrawler()
     visit = mock.Mock()
 
     @crawler.router.default_handler
     async def request_handler(context: PlaywrightCrawlingContext) -> None:
         visit(context.request.url)
-        await context.enqueue_links()
+        await context.enqueue_links(include=[Glob('https://crawlee.dev/docs/examples/**')])
 
     await crawler.run(requests)
 
     visited = {call[0][0] for call in visit.call_args_list}
 
-    assert visited == {
-        'https://crawlee.dev/',
-        'https://crawlee.dev/docs/guides/javascript-rendering',
-        'https://crawlee.dev/docs/guides/typescript-project',
-        'https://crawlee.dev/docs/guides/avoid-blocking',
-        'https://crawlee.dev/docs/guides/cheerio-crawler-guide',
-        'https://crawlee.dev/docs/guides/result-storage',
-        'https://crawlee.dev/docs/guides/proxy-management',
-        'https://crawlee.dev/api/core/class/AutoscaledPool',
-        'https://crawlee.dev/docs/guides/jsdom-crawler-guide',
-        'https://crawlee.dev/docs/guides/request-storage',
-        'https://crawlee.dev/api/utils',
-        'https://crawlee.dev/api/utils/namespace/social',
-        'https://crawlee.dev/docs/deployment/aws-cheerio',
-        'https://crawlee.dev/api/basic-crawler/interface/BasicCrawlerOptions',
-        'https://crawlee.dev/docs/deployment/gcp-cheerio',
-        'https://crawlee.dev/docs/quick-start',
-    }
+    assert len(visited) >= 10
+    assert all(url.startswith('https://crawlee.dev/docs/examples') for url in visited)